# БР1.2/Москалец Данила/homeworks/hw2/openapi.yaml
openapi: 3.0.0
info:
  title: Sangel API
  description: Маркетплейс охранных услуг
  version: 1.0.0
  contact:
    name: Москалец Данила Алексеевич
servers:
  - url: https://api.sangel.ru/v1
    description: Production server
  - url: http://localhost:8000/v1
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Pagination:
      type: object
      properties:
        total:
          type: integer
        offset:
          type: integer
        limit:
          type: integer
        has_more:
          type: boolean

    # User
...
    Category:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        is_published:
          type: boolean
...
    Service:
      type: object
      properties:
        id:
          type: integer
        company_id:
          type: integer
        company_title:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: decimal
        final_price:
          type: number
          format: decimal
        discount_percentage:
          type: integer
        is_published:
          type: boolean
        categories:
          type: array
          items:
            $ref: '#/components/schemas/Category'
        avg_rating:
          type: number
          format: float
    CreateServiceRequest:
      type: object
      required:
        - name
        - price
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        is_published:
          type: boolean
        category_ids:
          type: array
          items:
            type: integer
    UpdateServiceRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        is_published:
          type: boolean
        category_ids:
          type: array
          items:
            type: integer

    # Discount
...
    Request:
      type: object
      properties:
        id:
          type: integer
        service:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
            company:
              type: object
              properties:
                id:
                  type: integer
                title:
                  type: string
        user:
          type: object
          properties:
            id:
              type: integer
            first_name:
              type: string
            last_name:
              type: string
        status:
          type: string
          enum: [PENDING, ACCEPTED, REJECTED, CANCELLED]
        description:
          type: string
        reply:
          type: string
        created_at:
          type: string
          format: date-time
...
security:
  - bearerAuth: []
paths:
...
  /services:
    get:
      summary: Список услуг
      tags: [Services]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: company_id
          in: query
          schema:
            type: integer
        - name: category_id
          in: query
          schema:
            type: integer
        - name: price_min
          in: query
          schema:
            type: number
        - name: price_max
          in: query
          schema:
            type: number
        - name: is_published
          in: query
          schema:
            type: boolean
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [price, rating, created_at]
            default: created_at
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        200:
          description: Успешно
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Service'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
  /services/{id}:
    get:
      summary: Детальная информация об услуге
      tags: [Services]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Успешно
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Service'
        404:
          description: Услуга не найдена

    put:
      summary: Обновить услугу
      tags: [Services]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateServiceRequest'
      responses:
        200:
          description: Услуга обновлена
        403:
          description: Доступ запрещен
        404:
          description: Услуга не найдена

    delete:
      summary: Удалить услугу
      tags: [Services]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Услуга удалена
        403:
          description: Доступ запрещен
        404:
          description: Услуга не найдена
  /companies/{company_id}/services:
    post:
      summary: Создать услугу
      tags: [Services]
      parameters:
        - name: company_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateServiceRequest'
      responses:
        201:
          description: Услуга создана
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Service'
        403:
          description: Доступ запрещен
        404:
          description: Компания не найдена
  /services/{service_id}/requests:
    get:
      summary: Получить все заявки на услугу
      tags: [Services]
      parameters:
        - name: service_id
          in: path
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, ACCEPTED, REJECTED, CANCELLED]
      responses:
        200:
          description: Успешно
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Request'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        403:
          description: Доступ запрещен (не владелец компании)
        404:
          description: Услуга не найдена

  # ==================== DISCOUNTS ====================
...